+++ golang/metadata.rb
@@ -1,7 +1,0 @@
-name             'golang'
-maintainer       'timakin'
-maintainer_email 'timaki.st@gmail.com'
-license          'All rights reserved'
-description      'Installs/Configures golang'
-long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.0'