+++ goserver/CHANGELOG.md
@@ -1,9 +1,9 @@
-# golang CHANGELOG
+# goserver CHANGELOG
 
-This file is used to list changes made in each version of the golang cookbook.
+This file is used to list changes made in each version of the goserver cookbook.
 
 ## 0.1.0
-- [your_name] - Initial release of golang
+- [your_name] - Initial release of goserver
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.